Disable the pip cache to prevent a warning message
The following warning was showing up in the build output:

WARNING: The directory '/usr/local/airflow/.cache/pip' or its parent
directory is not owned or is not writable by the current user. The cache
has been disabled. Check the permissions and owner of that directory. If
executing pip with sudo, you may want sudo's -H flag.