Import Kubernetes certificates into truststore
This is required for the login with OpenShift to work. It allows
Airflow to connect to the openshift.default.svc.cluster.local and
oauth-openshift.apps.<cluster_domain> endpoints while executing
the OAuth authorization code flow.

Without trusting the Kubernetes certs, Airflow will refuse to
connect as it cannot verify the server TLS certificate.